commit 7 - añado notas y exporto notas en csv

+++ src/main/java/com/gestion/notasExamen/entity/Exam_StudentEntity.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -27,7 +29,7 @@ public class Exam_StudentEntity {
   private StudentEntity student;
 
   @Column(name="note")
-  private byte note;
+  private BigDecimal note;
 
 
 }
